Subsystem hamiltonian needs nuclear attraction per core
This means I had to dig into the ThIng code, which is not fun, and I made arrays of U matrices in CartInt and SphInt.

The energy is still the same, so now I can construct the subsystem Hamiltonian using the correct nuclear attraction elements.

One problem that still remains is how to construct the subsystem hamiltonian, probably the best thing is to keep a SphInt object in the SubSys class and "orthogonalize and project" it. Afterwards I can use this to construct the moleculer Hamiltonian.